Move filename class to new namespace
